Update lai Sua Xoa Song

+++ Project-Zing/src/main/java/com/codegym/repository/SongRepository.java
@@ -3,5 +3,5 @@ package com.codegym.repository;
 import com.codegym.model.Song;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface SongRepository extends PagingAndSortingRepository<Song, Integer> {
+public interface SongRepository extends PagingAndSortingRepository<Song, Long> {
 }
